final price process

+++ database/migrations/2024_02_05_032559_create_discount_matrix_table.php
@@ -17,7 +17,7 @@ class CreateDiscountMatrixTable extends Migration
             $table->id();
             $table->integer('from');
             $table->integer('to');
-            $table->integer('applied_discount');
+            $table->decimal('applied_discount', 8, 2);
             $table->timestamps();
         });
     }
